Revert unitEnum requirement
