--- lib/Interfaces/Broadcaster.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Underpin\Interfaces;
-
-
-namespace Underpin\Interfaces;
-
-use UnitEnum;
-
-interface Broadcaster {
-
-	function attach( UnitEnum $key, Observer $observer ): static;
-
-	function detach( UnitEnum $key, string $observer_id ): static;
-
-}
